+++ src/app/core/services/constant.service.ts
@@ -1,3 +1,0 @@
-import { InjectionToken } from '@angular/core';
-
-export const ConstantService = new InjectionToken<any>('constants value');