+++ src/app/core/services/constants.service.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const ConstantsService = new InjectionToken<any>('constants value');