+++ src/app/core/services/index.ts
@@ -1,4 +1,6 @@
 export * from './auth.service';
 export * from './local-storage.service';
 export * from './config-option.service.ts.service';
-export * from './constant.service';
+export * from './constants.service';
+export * from './generator.service';
+export * from './generator-factory.service';