+++ src/app/layout/components/about/about.component.ts
@@ -1,19 +1,27 @@
-import { Component, OnInit, Optional } from '@angular/core';
+import { Component, OnInit, Optional, Inject } from '@angular/core';
 
-import { СonfigOptionsService } from '../../../core/services';
+import {  } from '../../../core/services';
+import { ConstantsService, GeneratorS, GeneratorFactoryService, GeneratorService } from '../../../core/services';
 
 @Component({
   selector: 'app-about',
   templateUrl: './about.component.html',
-  styleUrls: ['./about.component.scss']
+  styleUrls: ['./about.component.scss'],
+  providers: [
+    GeneratorService,
+    { provide: GeneratorS, useFactory: GeneratorFactoryService(10), deps: [GeneratorService] }
+  ]
 })
 export class AboutComponent implements OnInit {
+  contentRand: string;
 
   constructor(
-    @Optional() private configService: СonfigOptionsService
+    @Optional() @Inject(ConstantsService) private constants: object,
+    @Optional() @Inject(GeneratorS) private generatorS: string
   ) { }
 
   ngOnInit() {
+    this.contentRand = `${this.generatorS}`;
+    console.log(this.constants ? `ConstantsService: ${JSON.stringify(this.constants)}` : 'ConstantsService not found.');
   }
-
 }